requirements.txt
absl-py==1.2.0
einops==0.4.1
h5py==3.7.0
keopscore==2.1
opt-einsum==3.3.0
pandas==1.4.2 
pytorch-wavelet 
PyWavelets==1.4.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
statsmodels==0.13.2
sympy==1.11.1
transformers==4.30.1
sktime==0.4.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
matplotlib==3.5.1
tensorboard==2.13.0
ipdb==0.13.13
tabulate==0.9.0
wandb==0.18.6
torch_geometric==2.3.0